Move new methods to geometry/optimization, per Russ' suggestion.
