+++ bindings/pydrake/planning/planning_py_trajectory_optimization.cc
@@ -462,40 +462,6 @@ void DefinePlanningTrajectoryOptimization(py::module m) {
             py::arg("trajectory"), cls_doc.NormalizeSegmentTimes.doc);
   }
 
-  {
-    using drake::geometry::optimization::ConvexSet;
-    m.def(
-        "PartitionConvexSet",
-        [](const ConvexSet& convex_set,
-            const std::vector<int>& continuous_revolute_joints,
-            const double epsilon) {
-          std::vector<copyable_unique_ptr<ConvexSet>> copyable_result =
-              PartitionConvexSet(
-                  convex_set, continuous_revolute_joints, epsilon);
-          std::vector<std::unique_ptr<ConvexSet>> result(
-              std::make_move_iterator(copyable_result.begin()),
-              std::make_move_iterator(copyable_result.end()));
-          return result;
-        },
-        py::arg("convex_set"), py::arg("continuous_revolute_joints"),
-        py::arg("epsilon") = 1e-5, doc.PartitionConvexSet.doc);
-    m.def(
-        "PartitionConvexSet",
-        [](const std::vector<ConvexSet*>& convex_sets,
-            const std::vector<int>& continuous_revolute_joints,
-            const double epsilon = 1e-5) {
-          std::vector<copyable_unique_ptr<ConvexSet>> copyable_result =
-              PartitionConvexSet(CloneConvexSets(convex_sets),
-                  continuous_revolute_joints, epsilon);
-          std::vector<std::unique_ptr<ConvexSet>> result(
-              std::make_move_iterator(copyable_result.begin()),
-              std::make_move_iterator(copyable_result.end()));
-          return result;
-        },
-        py::arg("convex_sets"), py::arg("continuous_revolute_joints"),
-        py::arg("epsilon") = 1e-5, doc.PartitionConvexSet.doc);
-  }
-
   m.def("GetContinuousRevoluteJointIndices", &GetContinuousRevoluteJointIndices,
       py::arg("plant"), doc.GetContinuousRevoluteJointIndices.doc);
 }
