+++ planning/trajectory_optimization/BUILD.bazel
@@ -109,6 +109,7 @@ drake_cc_library(
         "//common",
         "//common/trajectories:bezier_curve",
         "//common/trajectories:composite_trajectory",
+        "//geometry/optimization:geodesic_convexity",
         "//geometry/optimization:graph_of_convex_sets",
         "//multibody/plant",
         "//solvers:solve",
