+++ planning/trajectory_optimization/gcs_trajectory_optimization.cc
@@ -9,6 +9,7 @@
 #include "drake/common/scope_exit.h"
 #include "drake/common/symbolic/decompose.h"
 #include "drake/geometry/optimization/cartesian_product.h"
+#include "drake/geometry/optimization/geodesic_convexity.h"
 #include "drake/geometry/optimization/hpolyhedron.h"
 #include "drake/geometry/optimization/intersection.h"
 #include "drake/geometry/optimization/point.h"
@@ -32,13 +33,17 @@ using Eigen::MatrixXd;
 using Eigen::SparseMatrix;
 using Eigen::VectorXd;
 using geometry::optimization::CartesianProduct;
+using geometry::optimization::CheckIfSatisfiesConvexityRadius;
 using geometry::optimization::ConvexSet;
 using geometry::optimization::ConvexSets;
+using geometry::optimization::GetMinimumAndMaximumValueAlongDimension;
 using geometry::optimization::GraphOfConvexSets;
 using geometry::optimization::GraphOfConvexSetsOptions;
 using geometry::optimization::HPolyhedron;
 using geometry::optimization::Intersection;
+using geometry::optimization::PartitionConvexSet;
 using geometry::optimization::Point;
+using geometry::optimization::ThrowsForInvalidContinuousJointsList;
 using multibody::Joint;
 using multibody::JointIndex;
 using multibody::MultibodyPlant;
@@ -65,70 +70,6 @@ using VertexId = GraphOfConvexSets::VertexId;
 using EdgeId = GraphOfConvexSets::EdgeId;
 
 const double kInf = std::numeric_limits<double>::infinity();
-
-namespace {
-/* Computes the minimum and maximum values that can be attained along a certain
-dimension for a point constrained to lie within a convex set. */
-const std::pair<double, double> GetMinimumAndMaximumValueAlongDimension(
-    const ConvexSet& region, const int dimension) {
-  // TODO(cohnt) centralize this function in geometry/optimization, by making it
-  // a member of ConvexSet. Also potentially only compute the value once, and
-  // then return the pre-computed value instead of re-computing.
-  DRAKE_THROW_UNLESS(dimension >= 0 && dimension < region.ambient_dimension());
-  MathematicalProgram prog;
-  VectorXDecisionVariable y =
-      prog.NewContinuousVariables(region.ambient_dimension());
-  region.AddPointInSetConstraints(&prog, y);
-  VectorXd objective_vector = VectorXd::Zero(region.ambient_dimension());
-  objective_vector[dimension] = 1;
-  auto objective = prog.AddLinearCost(objective_vector, y);
-
-  auto result = Solve(prog);
-  if (!result.is_success()) {
-    throw std::runtime_error(
-        "GcsTrajectoryOptimization: Failed to compute lower bound of a convex "
-        "set!");
-  }
-  const double lower_bound = result.GetSolution(y)[dimension];
-
-  objective_vector[dimension] = -1;
-  objective.evaluator()->UpdateCoefficients(objective_vector);
-
-  result = Solve(prog);
-  if (!result.is_success()) {
-    throw std::runtime_error(
-        "GcsTrajectoryOptimization: Failed to compute upper bound of a convex "
-        "set!");
-  }
-  const double upper_bound = result.GetSolution(y)[dimension];
-
-  return {lower_bound, upper_bound};
-}
-
-/* Helper function to assert that a given list of continuous revolute joint
-indices satisfies the requirements for the constructor to
-GcsTrajectoryOptimization, as well as any static functions that may take in
-such a list. */
-void ThrowsForInvalidContinuousJointsList(
-    int num_positions, const std::vector<int>& continuous_revolute_joints) {
-  for (int i = 0; i < ssize(continuous_revolute_joints); ++i) {
-    // Make sure the unbounded revolute joints point to valid indices.
-    if (continuous_revolute_joints[i] < 0 ||
-        continuous_revolute_joints[i] >= num_positions) {
-      throw std::runtime_error(fmt::format(
-          "Each joint index in continuous_revolute_joints must lie in the "
-          "interval [0, {}). Joint index {} (located at {}) violates this.",
-          num_positions, continuous_revolute_joints[i], i));
-    }
-  }
-  std::unordered_set<int> comparison(continuous_revolute_joints.begin(),
-                                     continuous_revolute_joints.end());
-  if (comparison.size() != continuous_revolute_joints.size()) {
-    throw std::runtime_error(fmt::format(
-        "continuous_revolute_joints must not contain duplicate entries."));
-  }
-}
-}  // namespace
 
 Subgraph::Subgraph(
     const ConvexSets& regions,
@@ -1085,104 +1026,6 @@ GcsTrajectoryOptimization::NormalizeSegmentTimes(
     }
   }
   return CompositeTrajectory<double>(normalized_bezier_curves);
-}
-
-bool CheckIfSatisfiesConvexityRadius(
-    const ConvexSet& convex_set,
-    const std::vector<int>& continuous_revolute_joints) {
-  for (const int& j : continuous_revolute_joints) {
-    auto [min_value, max_value] =
-        GetMinimumAndMaximumValueAlongDimension(convex_set, j);
-    if (max_value - min_value >= M_PI) {
-      return false;
-    }
-  }
-  return true;
-}
-
-ConvexSets PartitionConvexSet(
-    const ConvexSet& convex_set,
-    const std::vector<int>& continuous_revolute_joints, const double epsilon) {
-  DRAKE_THROW_UNLESS(epsilon > 0);
-  DRAKE_THROW_UNLESS(epsilon < M_PI);
-  ThrowsForInvalidContinuousJointsList(convex_set.ambient_dimension(),
-                                       continuous_revolute_joints);
-  // Boundedness along dimensions corresponding to continuous revolute joints
-  // will be asserted by the GetMinimumAndMaximumValueAlongDimension calls
-  // below.
-
-  ConvexSets sets = MakeConvexSets(convex_set);
-  const double convexity_radius_step = M_PI - epsilon;
-  const int dim = convex_set.ambient_dimension();
-
-  std::vector<std::pair<double, double>> bbox(dim, std::make_pair(0, 0));
-
-  // We only populate the entries corresponding to continuous revolute joints,
-  // since the lower and upper limits of other joints aren't needed.
-  for (const int& i : continuous_revolute_joints) {
-    bbox[i] = GetMinimumAndMaximumValueAlongDimension(convex_set, i);
-  }
-  // The overall structure is to partition the set along each dimension
-  // corresponding to a continuous revolute joint. The partitioning is done by
-  // constructing axis-aligned bounding boxes, and intersecting them with the
-  // original set.
-  for (const int& i : continuous_revolute_joints) {
-    const double min_value = bbox[i].first;
-    const double max_value = bbox[i].second;
-    if (max_value - min_value >= M_PI) {
-      // Since we operate along one dimension at a time, all of the elements of
-      // sets will potentially violate the convexity radius, and thus must be
-      // partitioned. Sets which end up being empty are dropped. Because we are
-      // adding the new ConvexSets into sets, we track where the list of old
-      // sets ended, to prevent the infinite loop, and for later deletion.
-      const int j_max = sets.size();
-      for (int j = 0; j < j_max; ++j) {
-        for (double lower_bound = min_value; lower_bound < max_value;
-             lower_bound += convexity_radius_step - epsilon) {
-          Eigen::MatrixXd A = Eigen::MatrixXd::Zero(2, dim);
-          Eigen::VectorXd b = Eigen::VectorXd::Zero(2);
-          A(0, i) = 1;
-          b[0] = lower_bound + convexity_radius_step;
-          A(1, i) = -1;
-          b[1] = -lower_bound;
-          HPolyhedron chunk_bbox(A, b);
-          Intersection candidate_set(*sets[j], chunk_bbox);
-          if (!candidate_set.IsEmpty()) {
-            sets.push_back(copyable_unique_ptr<ConvexSet>(candidate_set));
-          }
-        }
-      }
-      // The sets which we have now partitioned are still stored between index 0
-      // and index j_max. We can now delete them, since we've replaced them with
-      // smaller pieces.
-      sets.erase(sets.begin(), sets.begin() + j_max);
-    }
-  }
-  return sets;
-}
-
-ConvexSets PartitionConvexSet(
-    const ConvexSets& convex_sets,
-    const std::vector<int>& continuous_revolute_joints, const double epsilon) {
-  DRAKE_THROW_UNLESS(convex_sets.size() > 0);
-  DRAKE_THROW_UNLESS(convex_sets[0] != nullptr);
-  ThrowsForInvalidContinuousJointsList(convex_sets[0]->ambient_dimension(),
-                                       continuous_revolute_joints);
-
-  int ambient_dimension = convex_sets[0]->ambient_dimension();
-  for (int i = 1; i < ssize(convex_sets); ++i) {
-    DRAKE_THROW_UNLESS(convex_sets[i] != nullptr);
-    DRAKE_THROW_UNLESS(convex_sets[i]->ambient_dimension() ==
-                       ambient_dimension);
-  }
-
-  ConvexSets sets;
-  for (int i = 0; i < ssize(convex_sets); ++i) {
-    ConvexSets new_sets = PartitionConvexSet(
-        *convex_sets[i], continuous_revolute_joints, epsilon);
-    sets.insert(sets.end(), new_sets.begin(), new_sets.end());
-  }
-  return sets;
 }
 
 std::vector<int> GetContinuousRevoluteJointIndices(
