+++ planning/trajectory_optimization/gcs_trajectory_optimization.h
@@ -547,56 +547,6 @@ class GcsTrajectoryOptimization final {
   std::vector<int> global_continuity_constraints_{};
 };
 
-/** Given a convex set, and a list of indices corresponding to continuous
-revolute joints, check whether or not the set satisfies the convexity radius.
-See §6.5.3 of "A Panoramic View of Riemannian Geometry", Marcel Berger for a
-general definition of convexity radius. When dealing with continuous revolute
-joints, respecting the convexity radius entails that each convex set has a width
-of stricty less than π along each dimension corresponding to a continuous
-revolute joint.
-@throws std::exception if continuous_revolute_joints has repeated entries, or if
-any entry is outside the interval [0, convex_set.ambient_dimension()). */
-bool CheckIfSatisfiesConvexityRadius(
-    const geometry::optimization::ConvexSet& convex_set,
-    const std::vector<int>& continuous_revolute_joints);
-
-/** Partitions a convex set into (smaller) convex sets whose union is the
-original set and that each respect the convexity radius as in
-CheckIfSatisfiesConvexityRadius. In practice, this is implemented as
-partitioning sets into pieces whose width is less than or equal to π-ϵ. Each
-entry in continuous_revolute_joints must be non-negative, less than
-num_positions, and unique.
-@param epsilon is the ϵ value used for the convexity radius inequality. The
-partitioned sets are made by intersecting convex_set with axis-aligned
-bounding boxes that respect the convexity radius. These boxes are made to
-overlap by ϵ radians along each dimension, for numerical purposes.
-@return the vector of convex sets that each respect convexity radius.
-@throws std::exception if ϵ <= 0 or ϵ >= π.
-@throws std::exception if the input convex set is unbounded along dimensions
-corresponding to continuous revolute joints.
-@throws std::exception if continuous_revolute_joints has repeated entries, or if
-any entry is outside the interval [0, convex_set.ambient_dimension()). */
-geometry::optimization::ConvexSets PartitionConvexSet(
-    const geometry::optimization::ConvexSet& convex_set,
-    const std::vector<int>& continuous_revolute_joints,
-    const double epsilon = 1e-5);
-
-/* Function overload to take in a list of convex sets, and partition all so as
-to respect the convexity radius. Every set must be bounded and have the same
-ambient dimension. Each entry in continuous_revolute_joints must be
-non-negative, less than num_positions, and unique.
-@throws std::exception unless every ConvexSet in convex_sets has the same
-ambient_dimension.
-@throws std::exception if ϵ <= 0 or ϵ >= π.
-@throws std::exception if any input convex set is unbounded along dimensions
-corresponding to continuous revolute joints.
-@throws std::exception if continuous_revolute_joints has repeated entries, or if
-any entry is outside the interval [0, ambient_dimension). */
-geometry::optimization::ConvexSets PartitionConvexSet(
-    const geometry::optimization::ConvexSets& convex_sets,
-    const std::vector<int>& continuous_revolute_joints,
-    const double epsilon = 1e-5);
-
 /** Returns a list of indices in the plant's generalized positions which
 correspond to a continuous revolute joint (a revolute joint with no joint
 limits). This includes the revolute component of a planar joint */
